+++ Server/UserService/UserServiceApplication/Services/UserService.cs
@@ -9,6 +9,8 @@ using UserServiceDataAccess.Dto;
 using Microsoft.AspNetCore.WebUtilities;
 using System.Text;
 using UserServiceDataAccess.Enums;
+using Hangfire;
+using Microsoft.Extensions.Logging;
 
 namespace UserServiceApplication.Services
 {
@@ -17,22 +19,26 @@ namespace UserServiceApplication.Services
                             IMapper mapper, 
                             IPasswordHasher passwordHasher, 
                             IValidator<User> validator,
-                            IEmailService emailService) : BaseService<User>(validator), IUserService
+                            IEmailService emailService, 
+                            ILogger<UserService> logger) : BaseService<User>(validator, logger), IUserService
     {
         private readonly IUserUnitOfWork _unitOfWork = unitOfWork;
         private readonly ITokenService _tokenService = tokenService;
         private readonly IEmailService _emailService = emailService;
         private readonly IMapper _mapper = mapper;
         private readonly IPasswordHasher _passwordHasher = passwordHasher;
-
-
+        private readonly ILogger<UserService> _logger = logger;
 
         public async Task<(string, string)> LoginAsync(LoginRequest loginRequest, CancellationToken cancellationToken)
-        {
+        {   
+            _logger.LogInformation("Login attempt started for {Email}", loginRequest.Email);
+
             var candidate = await _unitOfWork.UserRepository.GetByEmailAsync(loginRequest.Email, cancellationToken);
 
             if (candidate == null || !_passwordHasher.Verify(loginRequest.Password, candidate.Password))
             {
+                _logger.LogError("Login attempt failed for {Email}: Email or password is incorrect", loginRequest.Email);
+
                 throw new NotFoundException("Email or password is incorrect");
             }
 
@@ -40,20 +46,31 @@ namespace UserServiceApplication.Services
             var userClaims = _mapper.Map<UserClaimsDto>(candidate);
             var (accessToken, refreshToken) = await _tokenService.GenerateAuthTokensAsync(userClaims, cancellationToken);
 
-            await _unitOfWork.SaveAsync();
+            cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Login attempt completed successfully for {Email}", loginRequest.Email);
 
             return (accessToken, refreshToken);
         }
 
         public async Task<(string, string)> RegisterAsync(RegisterRequest registerRequest, CancellationToken cancellationToken)
         {
+            _logger.LogInformation("Register attempt started for {Email}", registerRequest.Email);
+
             var candidate = await _unitOfWork.UserRepository.GetByEmailAsync(registerRequest.Email, cancellationToken);
+
             if (candidate != null)
             {
-                throw new NotFoundException("User already exists");
+                _logger.LogError("Registration attempt failed for {Email}: user already exists", registerRequest.Email);
+
+                throw new ConflictException("User already exists");
             }
+
             registerRequest.Password = _passwordHasher.Generate(registerRequest.Password);
 
+            cancellationToken.ThrowIfCancellationRequested();
             var user = _mapper.Map<User>(registerRequest);
             Validate(user);
 
@@ -64,78 +81,137 @@ namespace UserServiceApplication.Services
             var userClaims = _mapper.Map<UserClaimsDto>(user);
             var (accessToken, refreshToken) = await _tokenService.GenerateAuthTokensAsync(userClaims, cancellationToken);
 
-            await _unitOfWork.SaveAsync();
+            cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Register attempt completed successfully for {Email}", registerRequest.Email);
 
             return (accessToken, refreshToken);
         }
 
         public async Task UpdateUserAsync(UpdateUserRequest updateUserRequest, CancellationToken cancellationToken)
         {
-            var candidate = await _unitOfWork.UserRepository.GetByIdAsync(updateUserRequest.Id, cancellationToken) ?? throw new NotFoundException("No user was found");
-            cancellationToken.ThrowIfCancellationRequested();
+            _logger.LogInformation("Update user attempt started for {Id}", updateUserRequest.Id);
 
+            var candidate = await _unitOfWork.UserRepository.GetByIdAsync(updateUserRequest.Id, cancellationToken);
+
+            if (candidate == null)
+            {
+                _logger.LogError("Update user attempt failed for {Id}: user not found", updateUserRequest.Id);
+
+                throw new NotFoundException("No user was found");
+            }
+
+            cancellationToken.ThrowIfCancellationRequested();
             var user = _mapper.Map(updateUserRequest, candidate);
             _unitOfWork.UserRepository.Update(user, cancellationToken);
 
             cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
 
-            await _unitOfWork.SaveAsync();
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Update user attempt completed successfully for {Id}", updateUserRequest.Id);
         }
 
         public async Task DeleteAsync(Guid id, CancellationToken cancellationToken)
         {
-            var candidate = await _unitOfWork.UserRepository.GetByIdAsync(id, cancellationToken) ?? throw new NotFoundException("No user found");
+            _logger.LogInformation("Delete user attempt started for {Id}", id);
+
+            var candidate = await _unitOfWork.UserRepository.GetByIdAsync(id, cancellationToken);
+
+            if (candidate == null)
+            {
+                _logger.LogError("Delete user attempt failed for {Id}: user not found", id);
+
+                throw new NotFoundException("No user was found");
+            }
+            cancellationToken.ThrowIfCancellationRequested();
             _unitOfWork.UserRepository.Delete(candidate, cancellationToken);
 
             cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
 
-            await _unitOfWork.SaveAsync();
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Delete user attempt completed successfully for {Id}", id);
         }
 
         public async Task<UserDto> GetUserByEmailAsync(string email, CancellationToken cancellationToken)
         {
-            var user = await _unitOfWork.UserRepository.GetByEmailAsync(email, cancellationToken) ?? throw new NotFoundException("No such user found");
+            var user = await _unitOfWork.UserRepository.GetByEmailAsync(email, cancellationToken);
+
+            if (user == null)
+            {
+                _logger.LogError("Get User by email attempt failed for {Email}: user not found", email);
+
+                throw new NotFoundException("No user was found");
+            }
+
             cancellationToken.ThrowIfCancellationRequested();
 
-            var userDto = _mapper.Map<UserDto>(user);
-
-            return userDto;
+            return _mapper.Map<UserDto>(user);
         }
 
         public async Task<UserDto> GetUserByIdAsync(Guid id, CancellationToken cancellationToken)
         {
-            var user = await _unitOfWork.UserRepository.GetByIdAsync(id, cancellationToken) ?? throw new NotFoundException("No such user found");
+            var user = await _unitOfWork.UserRepository.GetByIdAsync(id, cancellationToken);
+
+            if (user == null)
+            {
+                _logger.LogError("Get User by id attempt failed for {Id}: user not found", id);
+
+                throw new NotFoundException("No user was found");
+            }
+
             cancellationToken.ThrowIfCancellationRequested();
 
-            var userDto = _mapper.Map<UserDto>(user);
-
-            return userDto;
+            return _mapper.Map<UserDto>(user);
         }
 
         public async Task<string> ConfirmEmailSendAsync(string? accessToken, string callbackUrl, CancellationToken cancellationToken)
         {
-            var (confirmToken, email) = await _tokenService.GenerateTokenAndExtractEmailAsync(accessToken, TokenType.ConfirmEmail ,cancellationToken);
-            cancellationToken.ThrowIfCancellationRequested();
+            _logger.LogInformation("Confirm email send attempt started");
 
+            var (confirmToken, email) = await _tokenService.GenerateTokenAndExtractEmailAsync(accessToken, TokenType.ConfirmEmail ,cancellationToken);
+
+            cancellationToken.ThrowIfCancellationRequested();
             confirmToken = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(confirmToken));
-            SendEmail(email, confirmToken, "Confirm Email", callbackUrl, cancellationToken);
-            await _unitOfWork.SaveAsync();
+
+            BackgroundJob.Enqueue(() => SendEmail(email, confirmToken, "Confirm Email", callbackUrl, cancellationToken));
+
+            cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Confirm email send attempt completed successfully");
 
             return confirmToken;
         }
 
         public async Task<string> ConfirmEmailRecieveAsync(ConfirmEmailRequest confirmEmailRequest, CancellationToken cancellationToken)
         {
-            var candidate = await _unitOfWork.UserRepository.GetByEmailTrackingAsync(confirmEmailRequest.Email, cancellationToken) ?? throw new NotFoundException("No user found");
-            cancellationToken.ThrowIfCancellationRequested();
+            _logger.LogInformation("Confirm email recieve attempt started for {Email}", confirmEmailRequest.Email);
 
+            var candidate = await _unitOfWork.UserRepository.GetByEmailTrackingAsync(confirmEmailRequest.Email, cancellationToken);
+
+            if (candidate == null)
+            {
+                _logger.LogError("Confirm email recieve attempt failed for {Email}: user not found", confirmEmailRequest.Email);
+
+                throw new NotFoundException("No user was found");
+            }
+
+            cancellationToken.ThrowIfCancellationRequested();
             var confirmToken = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(confirmEmailRequest.Token));
             await _tokenService.FindAndDeleteTokenAsync(confirmToken, TokenType.ConfirmEmail, cancellationToken);
 
-            cancellationToken.ThrowIfCancellationRequested();
             candidate.IsEmailConfirmed = true;
 
-            await _unitOfWork.SaveAsync();
+            cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Confirm email recieve attempt completed successfully for {Email}", confirmEmailRequest.Email);
 
             return confirmToken;
         }
@@ -143,19 +219,36 @@ namespace UserServiceApplication.Services
 
         public async Task<string> ForgotPasswordAsync(string? accessToken, string callbackUrl, CancellationToken cancellationToken)
         {
+            _logger.LogInformation("Forgot password attempt started");
+
             var (resetToken, email) = await _tokenService.GenerateTokenAndExtractEmailAsync(accessToken, TokenType.ResetPassword, cancellationToken);
             cancellationToken.ThrowIfCancellationRequested();
 
             resetToken = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(resetToken));
-            SendEmail(email, resetToken, "Reset Password", callbackUrl, cancellationToken);
-            await _unitOfWork.SaveAsync();
+            BackgroundJob.Enqueue(() => SendEmail(email, resetToken, "Reset Password", callbackUrl, cancellationToken));
+
+            cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Forgot password attempt completed successfully");
 
             return resetToken;
         }
 
         public async Task<string> ResetPasswordAsync(ResetPasswordRequest resetPasswordRequest, CancellationToken cancellationToken)
         {
-            var candidate = await _unitOfWork.UserRepository.GetByEmailTrackingAsync(resetPasswordRequest.Email, cancellationToken) ?? throw new NotFoundException("No user found");
+            _logger.LogInformation("Reset password attempt started for {Email}", resetPasswordRequest.Email);
+
+            var candidate = await _unitOfWork.UserRepository.GetByEmailTrackingAsync(resetPasswordRequest.Email, cancellationToken);
+
+            if (candidate == null)
+            {
+                _logger.LogError("Reset password attempt failed for {Email}: user not found", resetPasswordRequest.Email);
+
+                throw new NotFoundException("No user was found");
+            }
+
             cancellationToken.ThrowIfCancellationRequested();
 
             var resetToken = Encoding.UTF8.GetString(WebEncoders.Base64UrlDecode(resetPasswordRequest.Token));
@@ -164,17 +257,23 @@ namespace UserServiceApplication.Services
             cancellationToken.ThrowIfCancellationRequested();
             candidate.Password = _passwordHasher.Generate(resetPasswordRequest.NewPassword);
 
-            await _unitOfWork.SaveAsync();
+            cancellationToken.ThrowIfCancellationRequested();
+            await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+            _logger.LogInformation("Changes to database were saved");
+            _logger.LogInformation("Reset password attempt completed successfully for {Email}", resetPasswordRequest.Email);
 
             return resetToken;
         }
 
-        private void SendEmail(string email, string token, string title, string callbackUrl, CancellationToken cancellationToken)
+        public void SendEmail(string email, string token, string title, string callbackUrl, CancellationToken cancellationToken)
         {
             callbackUrl += $"?{nameof(email)}={email}&{nameof(token)}={token}";
 
             _emailService.SendEmailAsync(email, title,
                 $"Please go through the link {callbackUrl}.", cancellationToken);
+
+            _logger.LogInformation("Sending email to {Email} with title {Title}", email, title);
         }
     }
 }
